src/requirements.txt
pywcmp==0.11.0
pywis-pubsub==0.8.0
pystac==1.13.0
py-multihash==2.0.1
